==> 01_participant_data.Rmd <==
---
title: "01_participant_data"
author: "Puvvula"
date: "2024-11-15"
output: pdf_document
---

```{r}
df_linkage <- read_csv("~/Documents/air_polln_rna_seq/crib_data/rna_seq_linkage.csv") |>
  clean_names()

crib_dat <- read_csv("~/Documents/air_polln_rna_seq/crib_data/crib_data.csv") |>
  clean_names()
```

#address linkage
```{r}
addr<- read_excel("~/Documents/peach_heat_metab/data/address_data/address_for_geo_cdng.xlsx") |> 
  clean_names()
addr_link<- read_excel("~/Documents/peach_heat_metab/data/address_data/mrn_link_for_addrs.xlsx") |> 
  clean_names()

address_df<- left_join(addr, addr_link, by="mrn")
```


#looking linkage
```{r}
df<- left_join(df_linkage, crib_dat, by= "subject_id")

filtered_address_df <- address_df |>
  filter(record_id %in% df$subject_id)

df_fin<- left_join(df, filtered_address_df, by=c("subject_id"="record_id")) |>
  select(c(1:3,6:12,14:18)) |>
  mutate(
    race = ifelse(is.na(race), ifelse(str_starts(birth_info, "AA"), "Black", "White"), race),
    neonatal_sex = ifelse(is.na(neonatal_sex), ifelse(str_ends(birth_info, "Female"), "Female", "Male"), neonatal_sex)
  )

write_csv(df_fin, "~/Documents/air_polln_rna_seq/crib_data/data_v_dec4.csv")
```

#adding additional available variables
```{r}
df_dec<- read_csv("~/Documents/air_polln_rna_seq/crib_data/data_v_dec4.csv") |>
  select(-c(4:6,9,10))

addl_vars<- read_csv("~/Documents/air_polln_rna_seq/crib_data/adding_missing_vars/date_ga_bmi.csv") |> clean_names()
addl_vars_2 <- read_csv("~/Documents/peach_heat_metab/metabolon_export/exp_cov_for_metabolon_dec3.csv") |> 
  clean_names() |>
  select(c(4:9,11)) |>
  select(1,5,6,7)

df<- left_join(df_dec, addl_vars, by= "subject_id") |>
  mutate(subject_id = as.character(subject_id)) |>
  left_join(addl_vars_2, by = c("subject_id" = "id"))

write_csv(df, "~/Documents/air_polln_rna_seq/crib_data/data_v_dec5.csv")
```

#read data received on Dec 11 - from rita and yu-chin
```{r}
rita <- readxl::read_xlsx("~/Documents/air_polln_rna_seq/crib_data/addn_data_dec11/crib_data_req 12-11-2024.xlsx") |>
  clean_names() |> mutate(delivery_date = format(as.Date(delivery_date), "%m-%d-%y"))
yu_chin <- readxl::read_xlsx("~/Documents/air_polln_rna_seq/crib_data/addn_data_dec11/Human placenta RNAseq Clinical Info_SPTB.xlsx") |> clean_names()
```

#filling missing data
```{r}
my_data<- read_csv("~/Documents/air_polln_rna_seq/crib_data/data_v_dec5.csv") |> 
  mutate(delivery_date = format(as.Date(delivery_date), "%m-%d-%y"))
  
  
  
final_df <- my_data |>
  mutate(
    delivery_date = ifelse(
      is.na(delivery_date),
      rita$delivery_date[match(subject_id, rita$subject_id)],
      delivery_date
    ),
    education = ifelse(
      is.na(education),
      rita$education[match(subject_id, rita$subject_id)],
      education
    ),
    tobacco_use_dur_preg = ifelse(
      is.na(tobacco_use_dur_preg),
      rita$tobacco_use_dur_preg[match(subject_id, rita$subject_id)],
      tobacco_use_dur_preg
    ),
    ga_at_delivery = ifelse(
      is.na(ga_at_delivery),
      yu_chin$ga_at_delivery[match(subject_id, yu_chin$subject_id)],
      ga_at_delivery
    ),
    delivery_mode = ifelse(
      is.na(delivery_mode),
      yu_chin$delivery_mode[match(subject_id, yu_chin$subject_id)],
      delivery_mode
    ),
    maternal_age = ifelse(
      is.na(maternal_age),
      yu_chin$maternal_age[match(subject_id, yu_chin$subject_id)],
      maternal_age
    ),
    bmi = ifelse(
      is.na(bmi),
      yu_chin$bmi[match(subject_id, yu_chin$subject_id)],
      bmi
    ),
    parity = ifelse(
      is.na(parity),
      yu_chin$parity[match(subject_id, yu_chin$subject_id)],
      parity
    )
  )

write_csv(final_df, "~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/crib_dat_final.csv")
```

#adding lat and long
```{r}
df<- read_csv("~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/crib_dat_final.csv") |>
  mutate(mrn = as.character(mrn))

metab_addrs_db<- read_csv("~/Documents/peach_heat_metab/data/pj_data/geocoded_addrs.csv") |>
  clean_names() |> select(c(2,8,9))

df_update<- left_join(df, metab_addrs_db, by = "mrn")

write_csv(df_update, "~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/crib_dat_final.csv")
```

#fill missing addressess
```{r}
missing_addrs <- read_csv("~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/geocoded.csv")

df<- read_csv("~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/crib_dat_final.csv") |>
  mutate(
  lat = ifelse(
    is.na(lat),
    missing_addrs$lat[match(subject_id, missing_addrs$subject_id)],
    lat
  ),
  long = ifelse(
    is.na(long),
    missing_addrs$long[match(subject_id, missing_addrs$subject_id)],
    long
  )
)

write_csv(df, "~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/archive/crib_dat_final.csv")

#formatting for analysis
#5 participants missing BMI imputed using ptb group specific median
#1 missing with parity imputed to parus

df_analysis<- df |>
  mutate(ptb = word(birth_info, 2),
         bmi = ifelse(
           is.na(bmi) & ptb == "Preterm", 
           median(bmi[ptb == "Preterm"], na.rm = TRUE), 
           bmi)
         ) |>
  select(-c(2,6:10)) |>
  mutate((delivery_date = mdy(delivery_date),
         delivery_mode = str_replace_all(delivery_mode, "C-?Section", "C-section"),
         parity = replace_na(parity, "parus"),education = case_when(
           education %in% c("Less than High School", "Completed High school", 
                            "Less than College Degree", 
                            "Unknown/ Patient declined to answer") ~ "less_than_bachelors",
           TRUE ~ "greater_than_bachelors"),
         across(3:15, ~ if (is.character(.)) as.factor(.) else .)
         )

save(df_analysis, file = "~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/df_analysis.rda")
```

#basic viz
```{r}
ggplot(df_analysis, aes(long, lat)) +
    borders("county", regions = c("Pennsylvania", "New Jersey", "Delaware")) + 
    geom_point(aes(color = factor(ptb)), size=1) +
    ggspatial::annotation_scale(plot_unit = "km", location= "br")+
    theme_void()+
    theme(legend.position = "bottom", legend.box = "horizontal", legend.justification = "center")+
    guides(color = guide_legend(override.aes = list(shape = 15, size = 5), title = NULL))

ggsave("~/Documents/air_polln_rna_seq/crib_data/data_clean_dec11_2024/mapping_participants.tiff",
       bg= "white",
       dpi=300)
```

#manuscript table
```{r}
df_mans_tbl<- df_analysis |> select(c(1:12))

# Numeric variables summary
numeric_summary <- df_mans_tbl %>%
  group_by(ptb) %>%
  summarise(across(
    where(is.numeric), 
    list(
      median = ~round(median(., na.rm = TRUE), 2),
      p25 = ~round(quantile(., probs = 0.25, na.rm = TRUE), 2),
      p75 = ~round(quantile(., probs = 0.75, na.rm = TRUE), 2)
    )
  )) %>%
  pivot_longer(
    cols = -ptb,
    names_to = c("variable", ".value"),
    names_pattern = "(.+)_(.+)"
  )

# Categorical variables summary
df_mans_tbl %>%
  select(-c(1,2)) |>
    tbl_summary(
        by = ptb,
        type = all_categorical() ~ "categorical",
        statistic = all_categorical() ~ "{n} ({p}%)",
        percent = "column",
        missing = "no"
    ) %>%
    add_overall(last = TRUE) %>%
    modify_header(label = "**Variable**") %>%
    modify_spanning_header(c("stat_1", "stat_2", "stat_0") ~ "**PTB Levels**")

#exposure values
cols_to_keep <- names(ap_exposures) %>% 
  str_subset("^(pm|bc)_\\d+") %>%          # Find pm_*/bc_* columns
  keep(~as.numeric(str_extract(., "\\d+")) < 142)

exp_summary <- ap_exposures |>
  filter(subject_id %in% df_mans_tbl$subject_id) |>
  select(
    subject_id,            # This line keeps the subject_id
    all_of(cols_to_keep)   # This keeps the filtered pm/bc columns
  ) |>
  left_join(df_mans_tbl %>% select(subject_id, ptb), by = "subject_id") |>
  group_by(subject_id, ptb
           ) |>
  summarise(
    pm_pct_days = round((sum(across(starts_with("pm_"), ~ sum(. > 9, na.rm = TRUE)))/141)*100, 2),
    bc_pct_days = round((sum(across(starts_with("bc_"), ~ sum(. > 1, na.rm = TRUE)))/141)*100, 2)
  )

exp_summary |>
    tbl_summary(
        by = ptb
    ) 
```

#supplemental analysis
```{r}
ext_analysis <- ap_exposures |>
  filter(subject_id %in% df_mans_tbl$subject_id) |>
  select(
    subject_id,            # This line keeps the subject_id
    all_of(cols_to_keep)   # This keeps the filtered pm/bc columns
  ) |>
  left_join(df_mans_tbl %>% select(subject_id, ptb,
                                   race, maternal_age, bmi, education, tobacco_use_dur_preg, parity
                                   ), by = "subject_id") |>
  group_by(subject_id, ptb,
           race, maternal_age, bmi, education, tobacco_use_dur_preg, parity
           ) |>
  #summarise(
   # pm_days = sum(across(starts_with("pm_"), ~ sum(. > 9, na.rm = TRUE))),
    #bc_days = sum(across(starts_with("bc_"), ~ sum(. > 1, na.rm = TRUE)))
  #)
  summarise(
    # For pm_days, count occurrences of values above 9 in blocks of 30 pm_ variables
    pm_days_20w = sum(across(starts_with("pm_"), ~ sum(. > 9, na.rm = TRUE))),
    pm_days_1 = sum(across(2:31, ~ sum(. > 9, na.rm = TRUE))),
    pm_days_2 = sum(across(32:61, ~ sum(. > 9, na.rm = TRUE))),
    pm_days_3 = sum(across(62:91, ~ sum(. > 9, na.rm = TRUE))),
    pm_days_4 = sum(across(92:121, ~ sum(. > 9, na.rm = TRUE))),
    pm_days_5 = sum(across(122:142, ~ sum(. > 9, na.rm = TRUE))),
    
    # Similarly for bc_days, count occurrences of values above 1 in blocks of 30 bc_ variables
    bc_days_20w = sum(across(starts_with("bc_"), ~ sum(. > 1, na.rm = TRUE))),
    bc_days_1 = sum(across(143:172, ~ sum(. > 1, na.rm = TRUE))),
    bc_days_2 = sum(across(173:202, ~ sum(. > 1, na.rm = TRUE))),
    bc_days_3 = sum(across(203:232, ~ sum(. > 1, na.rm = TRUE))),
    bc_days_4 = sum(across(233:263, ~ sum(. > 1, na.rm = TRUE))),
    bc_days_5 = sum(across(264:283, ~ sum(. > 1, na.rm = TRUE)))
  ) |>
  ungroup()

#export pm/bc anomaly counts
#exp_ap_count<- ext_analysis |> select(c(1,9:18))
#write_csv(exp_ap_count, "~/Documents/air_polln_rna_seq/crib_data/ap_anomaly_counts.csv")

ext_analysis$ptb <- relevel(ext_analysis$ptb, ref = "Term")

#logistic regression with count of days
logit<- glm(ptb ~ pm_days_5 +
              race + maternal_age + bmi + 
              education + tobacco_use_dur_preg + parity, 
            data = ext_analysis, family = "binomial")
exp(cbind(OR = coef(logit), confint(logit)))
```
